Add deployer client
deployer client can be used to deploy features in a CAPI cluster.
It has workers. Each worker runs in a separate goroutine and picks
a queued job.

Following queues are used:
- dirty contains all requests that have requested to configure a feature
and are currenlty waiting to be served;
- inProgress contains all request that are currently being served;
- jobQueue contains all requests that needs to be served;
- results contains results for processed request.

A new request is added to dirty and in jobQueue,unless already
present. If a result for such request was available, result is cleared.
When a worker picks a request, such request is moved from jobQueue to
inProgress.
When a request is completed, successfully or not, then:
- result is stored in results;
- request is removed from InProgress;
- if request is in dirty, it gets removed from there and added to
jobQueue.