ClusterSummary controller
ClusterSummary Reconciler deployes needed feature in CAPI clusters.

This reconciler watches for WorkloadRoles. When a WorkloadRole reference
is created/deleted or Spec updated, all the ClusterSummaries referencing
the modified WorkloadRole instance are queued for reconciliation.

ClusterSummaryReconciler has two internal maps used for above scope.
ClusterSummary references WorkloadRoles. When a WorkloadRole changes, all the ClusterSummaries referencing it need to be
reconciled. In order to achieve so, ClusterSummary reconciler could watch for WorkloadRoles. When a WorkloadRole spec changes,
find all the ClusterSummaries currently referencing it and reconcile those. Problem is no I/O should be present inside a MapFunc
(given a WorkloadRole, return all the ClusterSummary referencing such WorkloadRole).
In the MapFunc, if the list ClusterSummaries operation failed, we would be unable to retry or re-enqueue the ClusterSummaries
referencing the WorkloadRole that changed.
Instead the approach taken is following:
- when a ClusterSummary is reconciled, update the WorkloadRoleMap;
- in the MapFunc, given the WorkloadRole that changed, we can immeditaly get all the ClusterSummaries needing a reconciliation (by
using the WorkloadRoleMap);
- if a ClusterSummary is referencing a WorkloadRole but its reconciliation is still queued, when WorkloadRole changes, WorkloadRoleMap
won't have such ClusterSummary. This is not a problem as ClusterSummary reconciliation is already queued and will happen.

The ClusterSummaryMap is used to update WorkloadRoleMap. Consider following scenarios to understand the need:
1. ClusterSummary A references WorkloadRoles 1 and 2. When reconciled, WorkloadRoleMap will have 1 => A and 2 => A;
and ClusterSummaryMap A => 1,2
2. ClusterSummary A changes and now references WorkloadRole 1 only. We ned to remove the entry 2 => A in WorkloadRoleMap. But
when we reconcile ClusterSummary we have its current version we don't have its previous version. So we use ClusterSummaryMap (at this
point value stored here corresponds to reconciliation #1. We know currently ClusterSummary references WorkloadRole 1 only and looking
at ClusterSummaryMap we know it used to reference WorkloadRole 1 and 2. So we can remove 2 => A from WorkloadRoleMap. Only after this
update, we update ClusterSummaryMap (so new value will be A => 1)