ClusterSummary cleanup
If sync mod is set to continuos:
- when a WorkloadRole instance is not referenced anymore, corresponding
Role/ClusterRole is removed from the CAPI Cluster;
- when a ClusterSummary is deleted, before removing the finalizer,
remove all Roles/ClusterRoles deployed by such ClusterSummary instance
in the CAPI Cluster

Also, given same feature can be deployed or undeployed from a CAPI
Cluster, DeployerInterface methods requires an extra argurment, a bool
to indicate whether that is a cleanup or a provision request.

When a provision request happens, logic first checks whether a cleanup
is in progress. If so, provision request errors out.
When a cleanup request happens, logic first checks whether a provision
is in progress. If so, cleanup request errors out.
That's because if a request to cleanup and a request to provision the same
feature in the same CAPI cluster go in parallel, the outcome might be buggy
(either some policies become stale or go missing).