Add gatekeeper support
When deploying Gatekeeper, ConstraintTemplate must be installed first.
When removing stale Gatekeeper policies, ConstraintTemplate must be
removed last.

Policies deployed in a CAPI cluster, have followings:
1. ConfigLabelNamespace/ConfigLabelName labels, containing the info on
the ConfigMap containing such a policy in the control cluster;
2. OwnerReferences are set to contain all ClusterSummaries that caused
such a policy to be deployed.

1. is needed to handle scenarios like following:
- two different ClusterFeatures both match same CAPI Cluster;
- each ClusterFeature references a different ConfigMap, let's call those
ConfigMapA and ConfigMapB;
- Both ConfigMapA and ConfigMapB contains same policy, same
namespace/name, but different content.

In scenarios like this, policy can be deployed only once. Let's say
content in ConfigMapA was deployed. When system tries to deploy same
policy again (as part of content of ConfigMapB) system needs to detect
this misconfiguration and fail deployment notifying customer.

This is reason ConfigLabelNamespace/ConfigLabelName are added.

2. is needed to handle scenarios like following:
- two different ClusterFeature match same CAPI cluster;
- both ClusterFeature references same ConfigMap;
- if one of the ClusterFeature stops matching the CAPI cluster or does
not reference ConfigMap anymore, cleanup will happen;
- when such cleanup happens, none of the policies contained in the
ConfigMap can be removed, as there is yet another ClusterFeature that
wants those policies deployed.

This is reason ClusterSummaries are listed as OwnerReferences of each
deployed policy. That info is used as reference count.