Helm charts
ClusterFeatureSpec has a section in Spec to list all helm charts that
need to be managed in matching clusters.

Action indicates whether install or uninstall a chart.
If action is set to install and chart version changes, that is treated
as an upgrade/downgrade.

This PR also removes specific section for
Kyverno/Gatekeeper/Contour/Prometheus. Having the ability to point
directly to helm chart to install, above options are not needed anymore.

Here is an example of a ClusterFeatureSpec requesting two helm charts to
be installed in each all CAPI Clusters matching the ClusterFeature.

```
spec:
  ...
  helmCharts:
  - repositoryURL: https://kyverno.github.io/kyverno/
    repositoryName: kyverno
    chartName: kyverno/kyverno
    chartVersion: v2.5.3
    releaseName: kyverno-latest
    releaseNamespace: kyverno
    helmChartAction: Install
  - repositoryURL: https://helm.nginx.com/stable
    repositoryName: nginx-stable
    chartName: nginx-stable/nginx-ingress
    chartVersion: 0.14.0
    releaseName: nginx-latest
    releaseNamespace: nginx
    helmChartAction: Install
```

Current limitations:
- Removing an entry from a ClusterFeatureSpec.HelmCharts won't uninstall
the corresponding chart. Current work-around: either set action to
Uninstall or delete the entire ClusterFeature
- If two ClusterFeatures match same cluster and try to manage same
release, such a conflict won't be detcted. No workaround for this.