Support referencing secrets
ClusterFeature can reference both Secrets and ConfigMaps.

When the content of a reference Secret changes, ClusterSummaries
referencing it are queued for reconciliation. Logic to detect which
ClusterSummaries to reconcile is the same used for ConfigMaps and relies
on internal in memory maps.

This PR adds needed logic along with uts and fv.