Referenced resources
When referencing a ConfigMap/Secret namespace can be set of left empty.

If namespace is set, it uniquely identifies a ConfigMap/Secret. The
content of such resource is deployed in each workload cluster matching
the ClusterProfile.

If namespace is left empty, at time of deployment, Sveltos will look for
a resource with provided name in the cluster namespace.
For instance setting:

PolicyRefs:
- namespace: ""
  name: foo
  kind: ConfigMap

will have Sveltos use:
1. ConfigMap foo in namespace A when deploying resources for a cluster
in namespace A;
2. ConfigMap foo in namespace B when deploying resources for a cluster
in namespace B.

This PR adds code, unit tests and functional verification test.