Deploy resources in mgmt clusters
Sveltos can be configured to deploy resources in the management cluster
as well.
An example of when this is needed is autoscaler with clusterAPI powered
clusters. Autoscaler requires a ServiceAccount to be created in the
management cluster and it requires those information to be passed to the
autoscaler instance running in the management cluster.

In the following example Sveltos is instructed to:
1. deploy content of the ConfigMap serviceaccount-autoscaler instance in
the management cluster (deploymentType: Local);
2. deploy content of the ConfigMap ecret-autoscaler instance in the
managed cluster (deploymentType: Remote)

```
apiVersion: config.projectsveltos.io/v1alpha1
kind: ClusterProfile
metadata:
  name: deploy-resources
spec:
  clusterSelector: env=fv
  policyRefs:
  - deploymentType: Local
    kind: ConfigMap
    name: serviceaccount-autoscaler
    namespace: default
  - deploymentType: Remote
    kind: ConfigMap
    name: secret-autoscaler
    namespace: default
```

Before configuring Sveltos to deploy resources in the management
cluster, Sveltos needs to be granted proper permissions. In order to do
so, modify the ClusterRole named addon-manager-role-extra.
Sveltos addon-manager ServiceAccount is already tied to that
ClusterRole.

This features is fully integrated with Sveltos multi-tenancy solution.
If a tenant admin creates a ClusterProfile with "deploymentType: Local"
Sveltos will only deploy it if the serviceAccount representing the
tenant has the proper permissions in the management cluster.

The resources to be deployed by Sveltos can come in the form of a
template.
Before this PR, few instances (Cluster, infrastructure provider,
SveltosCluster and few others were available. With this PR Sveltos can
be instructed to refer to any resource.

Following example Sveltos is instructed to fetch Secret instance named
autoscale in the cluster namespace. This resource is made available
during template instantiation and it can be referenced to as
AutoscalerSecret.

```
apiVersion: config.projectsveltos.io/v1alpha1
kind: ClusterProfile
metadata:
  name: deploy-resources
spec:
  clusterSelector: env=fv
  templateResourceRef:
  - resource:
      kind: Secret
      name: autoscaler
    identifier: AutoscalerSecret
  policyRefs:
  - deploymentType: Remote
    kind: ConfigMap
    name: secret-autoscaler
    namespace: default
---
apiVersion: v1
kind: ConfigMap
metadata:
  name: secret-autoscaler
  namespace: default
  annotations:
    projectsveltos.io/template: "true"
data:
  secret.yaml: |
    apiVersion: v1
    kind: Secret
    metadata:
      name: autoscaler
      namespace: {{ (index .MgtmResources "AutoscalerSecret").metadata.namespace }}
    data:
      token: {{ (index .MgtmResources "AutoscalerSecret").data.token }}
      ca.crt: {{ $data:=(index .MgtmResources "AutoscalerSecret").data }} {{ (index $data "ca.crt") }}
```