Always set DeployedGroupVersionKind
Cleanup of deployed resources needs to be aware of the GVKs deployed by
addon-manager.
If a subset of resources is deployed and then an error happens, before
this PR DeployedGroupVersionKind was not set. This causes addon-manager
to leave stale resource (on delete, given DeployedGroupVersionKind was
not set addon-manager used to clean up nothing).

This PR fixes this behavior by always setting DeployedGroupVersionKind.