Make management referenced resource name template
When referencing resources present in the management cluster allow
name to be a function of cluster info: cluster namespace, name and type.

For instance Sveltos might be instructed to create:
1.crossplane resources which end creating a google storage bucket for each
matching cluster
2. take bucket information and then pass to resources in the managed cluster
so to be used

In such a case, we want the crossplane Bucket resource to have a unique name
per cluster. And we want Sveltos to know what that name is.

```
apiVersion: config.projectsveltos.io/v1alpha1
kind: ClusterProfile
metadata:
  name: deploy-resources
spec:
  clusterSelector: env=fv
  templateResourceRefs:
  - resource:
      apiVersion: storage.cp.upbound.io/v1beta1
      kind: Bucket
      name: bucket-{{ .ClusterNamespace }}-{{ .ClusterName }}
    identifier: Bucket
  policyRefs:
  - deploymentType: Local
    kind: ConfigMap
    name: bucket
    namespace: default
---
apiVersion: v1
kind: ConfigMap
metadata:
  name: bucket
  namespace: default
  annotations:
    projectsveltos.io/template: "true"
data:
  bucket.yaml: |
    apiVersion: storage.gcp.upbound.io/v1beta1
    kind: Bucket
    metadata:
     name: bucket-{{ .Cluster.metadata.namespace }}-{{ .Cluster.metadata.name }}
     labels:
       docs.crossplane.io/example: provider-gcp
       clustername: {{ .Cluster.metadata.name }}
       clusternamespace: {{ .Cluster.metadata.namespace }}
    spec:
      forProvider:
        location: US
      providerConfigRef:
        name: default
```