Start drift-detection-manager in the management cluster
Addon-controller has new command line option: `agent-in-mgmt-cluster` when set
addon-controller will deploy drift-detection-manager in the management cluster.
Resources deployed in the management cluster are:
- service
- deployment

ResourceSummary CRD and instances will still be created in the managed cluster
as the drift-detection-manager will connect to the managed cluster.

When cluster is deleted, any drift-detection-manager resource created in the
management cluster will be removed