Add extra annotation
When deploying resources in the management add an extra annotation
to indicate the clusterSummary instance, and so the managed cluster,
for which the resource was created.

Resources deployed by Projectsveltos have ClusterProfile as OwnerReference(s).
In managed cluster that is enough to identify stale resources (resources deployed
by Projectsveltos which now needs to be removed). OwnerReference is not enough
for resources deployed by Projectsveltos in the management cluster.
For instance, a ClusterProfile:
- matching N managed clusters
- requested to create a crossplane Bucket instabce

will create N Bucket instances in the management cluster. One per matching managed
cluster.

If a managed cluster stops being a match, resources deployed for the managed cluster
needs to be removed. This include the Bucket instance deployed in the management cluster.

Projectsveltos logic used to:
1. query all Bucket instances deployed by Projectsveltos;
2. if the OwnerReference was the ClusterProfile, which is now cleaning, remove
ClusterProfile as OwnerReference and if no other OwnerReference was present, delete the
resource

Above logic though, cleaned also up all the Bucket instances for the other managed clusters.

This PR adds new annotation for resources deployed by Projectsveltos in the management
cluster indicating the mananged cluster for which resource was created.
This allows to fix above logic by simply adding an extra check:
- when cleaning up stale resources deployed by Projectsveltos in the management cluster
for a given managed cluster, this new annotation has to match, otherwise resources are
ignored.