Drift-detection logic
When ClusterSummary in mode ContinuousWithDriftDetection is about to
deploy resources, stop drift detection for such resources. This means:

- for helm resource, set Spec.ChartResources to empty
- for raw YAML/JSON, set Spec.Resources to empty
- for kustomize, set Spec.KustomizeResources to empty

This is done to avoid un-necessary reconciliation. If Sveltos is about
to redeploy those resources, very likely some of those will be modified.
By updating ResourceSummary instance before re-deploying those resources,
we avoid drift-detection from considering such updates (done by Sveltos)
as configuration drift.

Code already re-updates ResourceSummary after resources are re-deployed.

This PR also fix an issue where Sveltos reports an issue when no ReloaderReport
or ResourceSummary are present in the managed cluster. That is not an error
if those CRDs were not installed. Those CRDs are only installed in a cluster
when there is at least one matching ClusterProfile with Spec.Reloader field set
to true or mode set to ContinuousWithDriftDetection respectively.