Missing references
ClusterProfile can reference resources in both PolicyRefs and
KustomizationRef sections.

Before this PR, if a referenced resource was missing, addon-controller
simply reported in the logs. But feature was marked as Provisioned assuming
no other errors.

This used to create confusion as a simple typo would cause addon-controller
to mark feature as Provisioned while nothing was provisioned.

This PR changes this behavior.
If a referenced resource is missing, addon-controller will report the issue.
Here is a ClusterSummary status

```
  - failureMessage: 'Referenced resource: ConfigMap default/contour-gateway does not
      exist'
    featureID: Resources
    hash: riDS1DlksBGuKMyiCc03AysQHhWgjAtm3+amnwFRpbQ=
    lastAppliedTime: "2024-01-12T10:21:07Z"
    status: FailedNonRetriable
```

The error is marked as non retriable so nothing will happen till a change,
including creating the missing referenced resource, is made.