PolicyRefs: referenced ConfigMap/Secret with multiple resources
When ClusterProfile/Profile references a ConfigMap or a Secret
whose data section contains multiple resources, Sveltos tries to
deploy each one of them.

Before this PR, if a ConfigMap contained an entry in its Data section
with following resources:

1. deployment
2. service
3. secret

If deploying the deployment succeeded, but deploying service failed,
Sveltos did not update the
ClusterSummary.Status.featureSummaries.deployedGroupVersionKind

that field is used by Sveltos to understand which resources to cleanup
when a cleanup is needed.

So if in this state, ClusterProfile was deleted or the cluster stopped
being a match, the deployed Deployment instance was never removed from
the managed cluster.

This PR fixes that by always reporting the GVKs of the resources deployed
even an error happens mid way.