Improve conflict handling situation
In following situation:
1. profile 1 (tier 100) is currently deploying a resource
2. profile 2 (tier 50) wants to deploy same resource. A conflict
is detected but because of the tier values, profile 2 wins the
conflict.

Before this PR the logic was to requeue profile 1.

Problem with above logic is that profile 1 is already queued and
is deploying the resource. So when profile 1 reconciles again it
detects:

- nothing changed in its configuration
- a result is available in deployer and it is marked as successfull (
the deployment completed before profile 2 took over)

and so it marks itself as successfull.

This PR fixes this scenario. When profile 2 wins the conflict, before
requeuing profile 1, it checks if there is a job currently queueud for
profile 1. If there is, it returns an error.

If no job is queued, this PR also clears any potential result available
for profile 1.

The error will cause profile 2 to reconcile again later and at that
point requeuing profile 1 is good enough.