Watch for resources listed in TemplateResourceRefs
Profiles can use resources in the management cluster to instantiate
the templates representing the add-ons/applications to deploy in the
matching managed clusters.

Those resources are listed in the TemplateResourceRefs.

When one of those resources change, corresponding ClusterSummary
instances need to be requeued again. Templates need to be re-instantiated
again.

This PR starts and stop watchers for TemplateResourceRefs.