Introduce DeployedGVKs in ClusterSummary Status
ClusterSummary status is updated by:
1. ClusterSummary reconciler
2. worker deploying resources in the managed cluster

Point 2 only updates list of deployed Group/Version/Kinds.
That information is needed when a ClusterSummary is deleted to
find and remove resources deployed by such ClusterSummary instance.

This information is added for both KustomizationRefs and PolicyRefs.

Before this PR this information was stored in FeatureSummaries.
Because FeatureSummaries is also updated by point 1, there was a
condition were deployed GVKs was getting reset to empty thus incorrectly
leaving resources in the managed cluster when ClusterSummary was deleted.

This was a race condition:
- ClusterSummary reconciler starts
- worker deploying resources for this ClusterSummary ends (and stores
deployed GVKs)
- ClusterSummary reconciler ends and updates the status

Moving deployed GVK to a different section of Status which is only
updated by worker resolves this race condition.

DeployedGVKs is still not removed from FeatureSummaries.
This PR does not set DeployedGVKs in FeatureSummaries anymore but
still uses whatever values was set there when undeploying.
This is done to smoothly transition.

DeployedGVKs will be removed from FeatureSummaries in a few releases.