TemplateResourceRefs: name
TemplateResourceRefs can be expressed as a template and instantiated
at run-time using matching cluster namespace/name.

Before this PR:
- .ClusterNamespace represented the cluster namespace
- .ClusterName represented the cluster name

This is off with rest of templating where:
- .Cluster.metadata.namespace represents the cluster namespace
- .Cluster.metadata.name represents the cluster name

This PR allows .Cluster.metadata.namespace and .Cluster.metadata.name
to be also used. This will be the documented way.

In order not to break existing deployment, .ClusterNamespace and
.ClusterName can still be used.