(feat) Referenced resource names can be expressed as templates
For following fields:

- PolicyRef
- KustomizationRef
- ValueFrom

name can be expressed as a template and instantiate using

- cluster namespace: .Cluster.metadata.namespace
- cluster name: .Cluster.metadata.name
- cluster type: .Cluster.kind

It was already possible to leave namespace empty, in which case
cluster namespace was used.

Sveltos will instantiate name at run-time. This allows different
resources to be fetched for different cluster, all while using a
single ClusterProfile/Profile