Drift-detection-manager configuration
Sometimes it is useful to customize drift-detection-manager deployment.
Addon-controller accepts a new arg: `drift-detection-config`

That is the name of the ConfigMap in the `projectsveltos` namespace
that contains patches that will be applied to drift-detection-manager
before it gets deployed in the managed cluster.

For instance, creating this ConfigMap

```yaml
apiVersion: v1
data:
  patch: |-
    apiVersion: apps/v1
    kind: Deployment
    metadata:
      name: drift-detection-manager
    spec:
      template:
        spec:
          containers:
          - name: manager
            image: projectsveltos/drift-detection-manager:dev
            resources:
              requests:
                memory: 256Mi
            securityContext:
              readOnlyRootFilesystem: true
kind: ConfigMap
metadata:
  name: drift-detection
  namespace: projectsveltos
```

and setting the following extra argurment to addon-controller deployment

```
      - args:
        ...
        - --drift-detection-config=drift-detection
```

the drift-detection-manager will be deployed in each managed cluster with
256Mi for request memory and image `projectsveltos/drift-detection-manager:dev`