(feat) Rolling updates and paused cluster
When the `Spec.MaxUpdate` profile is set, only a specific number of clusterSummaries
can be updated at a time. Once these updates are successful, Sveltos selects additional
clusters for updating.

Paused clusters are excluded from this process.
If a paused cluster is accidentally selected, it will not be updated (cause it is set to paused)
and will prevent any other non paused clusters from being updated.