(feat) ClusterSet/Set: look at SveltosCluster ConnectionStatus
Sveltos periodically attempts to connect to managed SveltosClusters.
Upon successful initial connection, the `Ready` status is set to
true and remains unchanged.

To indicate connection health, this PR introduces a `ConnectionStatus`
field within the Status. This field can be either `Healthy` or `Down`.
The status transitions to `Down` after a configurable number of consecutive
connection failures, which defaults to three.

This PR changes ClusterSet/Set to only pick clusters that are:

1. Ready (it was already the case)
2. With `Healthy` ConnectionStatus