(bug) when cluster is marked as deleted avoid triggering cleanup
Clusters marked as deleted are scheduled for removal once their finalizer is removed.
Therefore, initiating cleanup while a cluster is in this state is redundant.

Prior to this change, Sveltos would trigger cleanup for a cluster instance even
if it was marked as deleted. However, if Sveltos invoked the Helm SDK as part
of the cleanup and the Helm uninstall client encountered an unreachable cluster
api-server, the Helm SDK would become stuck.

To prevent this issue and avoid unnecessary feature cleanup for clusters marked
for deletion, this PR has been implemented.