(bug) fix incorrect message about comptaibility check error
Before this PR this log was continuously printed in the add-on controller:

```
"failed to collect ResourceSummaries from cluster: mgmt/mgmt compatibility checks failed"
```

reason was the following:

1. ResourceSummary CRD is installed in the management cluster when Sveltos is deployed.
While in other cluster that is deployed only when drift-detection-manager is deployed.
2. So while fetching the ResourceSummary from the management cluster, ResourceSummary
CRD was found but then the ConfigMap with drift-detection-manager was not (obviously
because drift-detection-manager was not deployed).

This PR fixes this by checking the presence of the drift-detection-manager