Revert "(bug) when cluster is marked as deleted avoid triggering clea…
Sveltos deploys resources in the management cluster when the `deploymenType`
is set to `Local`.
Even if a cluster is marked as deleted, Sveltos will still need to remove
all resources deployed in the management cluster for that cluster.
So this PR lets Sveltos undeploy even for clusters marked as deleted.

If the managed cluster API server is unreachable after the Helm SDK
uninstall is triggered, the Helm SDK will get stuck. Helm charts are
only deployed in the matching cluster, so Sveltos will skip calling
Helm uninstall for any deleted cluster.

This reverts commit ce9e7f6.

Fixes #719