(bug) when deploymentType is Local, validate namespace
A Profile using deploymentType: Local, can only deploy resources
in the same namespace.

Previously, Sveltos used the following format for the OwnerReferences
name for resources it deployed due to a Profile:

```
Profile.namespace/Profile.name
```

However, a Profile with this exact name doesn't exist within the
management cluster. As a result, the Kubernetes controller would
immediately remove the deployed resource.

This behavior occurred because:

```
A namespaced owner must exist in the same namespace as the dependent.
If it does not, the owner reference is treated as absent, and the
dependent is subject to deletion once all owners are verified absent
```

Given this PR ensures that a Profile only creates resources in the management
cluster in the same namespace, now only Profile.name is set in the
OwnerReference.name