(feat) add `copy` function
It is common to request Sveltos to fetch a resource in the management
cluster and copy it over to managed clusters.

Here is an example where we want Sveltos to fetch the Secret
__eso-vault-secret-sync__ in the __default__ namespace and copy
it over all production clusters.

The `copy` function requires the ID of the resource to copy.
It resets the UUID and the ResourceVersion field, and copy
everything else.

```
apiVersion: config.projectsveltos.io/v1beta1
kind: ClusterProfile
metadata:
  name: sync-secrets
spec:
  clusterSelector:
    matchLabels:
      env: prod
  templateResourceRefs:
  - resource:
     apiVersion: v1
     kind:  Secret
     name: eso-vault-secret-sync
     namespace: default
    identifier: ExternalSecret
  policyRefs:
    - kind: ConfigMap
      name: info
      namespace: default
---
apiVersion: v1
kind: ConfigMap
metadata:
  name: info
  namespace: default
  annotations:
    projectsveltos.io/template: "true"
data:
  secret.yaml: |-
    {{ (copy "ExternalSecret") }}
```