(feat) add more template functions
- getField: return a field of a resource.
Requires the identifier of the resource fetched via TemplateResourceRefs and the field
For instance to get the Spec.Replicas of a Deployment.
```
{{ $replicasValue := getField "Deployment" "spec.replicas" }}
```

- removeField: removes a field from a resource
Requires the identifier of the resource fetched via TemplateResourceRefs and the field
For instance to remove Spec.Replicas of a Deployment
```
{{ removeField "Deployment" "spec.replicas" }}
```

- setField: set a field to a specific value
Requires the identifier of the resource fetched via TemplateResourceRefs, the field and new value
For instance to update Spec.Replicas of a Deployment
```
{{ setField "Deployment" "spec.replicas" (int64 7) }}
```

- The functions chainRemoveField chainSetField can be used to pipe more modifications
one after another.
For instance to modify Spec.Replicas, Namespace, ServiceAccount and Paused field
```
{{ $depl := (getResource "Deployment") }}
{{ $depl := (chainSetField $depl "spec.replicas" (int64 5) ) }}
{{ $depl := (chainSetField $depl "metadata.namespace" .Cluster.metadata.namespace ) }}
{{ $depl := (chainSetField $depl "spec.template.spec.serviceAccountName" "default" ) }}
{{ $depl := (chainSetField $depl "spec.paused" true ) }}
{{ toYaml $depl }}
```