(feat) CAPI clusters
addon-controller can be passed a new arg: `capi-onboard-annotation`.
It allows administrators to specify a custom annotation key.
When this annotation is set, Sveltos will only manage and apply
add-ons and applications to CAPI clusters that have this specific
annotation with any value.

CAPI clusters without this annotation will be ignored by the addon-controller.