Auto deploy dependencie: enable automatic deployment of prerequisite profile
Profile instances can declare dependencies on other profiles via the DependsOn field,
establishing a directed acyclic graph (DAG) of dependencies. Before this PR, it was the
administrator's responsibility to ensure that both the dependent and prerequisite
profiles target the same set of managed clusters through matching cluster selectors.

With this PR, Sveltos will automatically resolve and deploy the prerequisite profiles
specified in the DependsOn field. Sveltos will analyze the dependency graph, identify
the required prerequisite profiles, and ensure they are deployed to the same clusters.

This introduce a behavior change.
Set the addon-controller arg `auto-deploy-dependencies` to disable it.