(feat) HelmCharts: integration with Flux sources
Before this PR, the RepositoryURL was the URL helm chart repository
It could point to a traditional HTTP-based repository or an OCI registry.
Examples:
   - HTTP Repository: https://charts.example.com/stable
   - OCI Registry: oci://registry.example.com/namespace/charts
   - GitHub Pages: https://<username>.github.io/<repository>/
   - Any other valid URL where Helm charts are hosted.

With this PR, alternatively, it can be a Flux source reference:
   <flux source kind>://<flux source namespace>/<flux source name>/<path>

Where:
   - <flux source kind> is the type of Flux source (gitrepository, ocirepository, bucket).
   - <flux source namespace> and <flux source name> are the namespace and name of the Flux Source in the management cluster.
   - <path> is the path within the source repository where the Helm charts are located.

For instance:
   gitrepository://flux-system/flux-system/charts/projectsveltos

Assuming there is a Flux GitRepository named 'flux-system' in the 'flux-system' namespace
syncing 'https://github.com/projectsveltos/helm-charts.git/', and the Helm charts are
located in the 'charts/projectsveltos' directory of that repository.