(bug) drift detection compatibility check
Sveltos used to check the existance of ResourceSummary CRD in a managed
cluster as an indication drift-detection-manager was deployed in such
cluster.
Only after this check Sveltos was verifying the compatibility checks.

When running in agentless mode, no Sveltos CRDs are deployed anymore to
managed cluster as all resources used for drift detection and more are
all in the management cluster.

So this PR changes this check. When running in agentless mode, Sveltos
fetches all drift detection deployments(all in the management cluster)
and creates a list of clusters with drift detection enabled.
Then, compatibility checks is run only for those clusters.